+++ cmake/cora.cmake
@@ -8,6 +8,7 @@ set(cora_sources
     ${cora_source_dir}/commands/command.cpp
     ${cora_source_dir}/commands/command_interpreter.cpp
     ${cora_source_dir}/commands/import_command.cpp
+    ${cora_source_dir}/commands/find_command.cpp
     ${cora_source_dir}/commands/info_command.cpp
     ${cora_source_dir}/commands/help_command.cpp
     ${cora_source_dir}/commands/list_command.cpp