+++ src/commands/find_command.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "commands/command.hpp"
+
+class FindCommand : public Command {
+    public:
+        FindCommand();
+
+        virtual ~FindCommand();
+
+        virtual void execute(const std::vector<std::string>& args) override;
+};